+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 -------------------
 
 - Fixes/enhancements: #35, #37
+- Implemented ability to run buildings flood risk analysis (#36).
 
 
 0.3.5 (2024-9-12)